requirements.txt
litellm
pydantic
pandas
python-dotenv
google-generativeai
sqlite-database
